requirements.txt
moviepy
opencv-python
pika
Pillow
Flask
python-dotenv
